Cleanup, finish NLP processor
Cleaned up a few areas and improved the cascaded finite-state
transducer for natural language processing

+++ nltk/go.sh
@@ -1,25 +1,32 @@
 #!/usr/bin/env sh
 
-echo
-date
-echo "Running pre-pre-process a (remove useless lines) ..."
-./prepreprocess_a.py hddata_a.csv > hddata_b.csv
+if [ -e "hddata_a.csv" ]; then
+	echo
+	date
+	echo "Running pre-pre-process a (remove useless lines) ..."
+	./prepreprocess_a.py hddata_a.csv > hddata_b.csv
+fi
 
-echo
-date
-echo "Running pre-pre-process b (remove usernames, etc.) ..."
-./prepreprocess_b.py hddata_b.csv > hddata_c.csv
-rm hddata_b.csv
+if [ -e "hddata_b.csv" ]; then
+	echo
+	date
+	echo "Running pre-pre-process b (remove usernames, etc.) ..."
+	./prepreprocess_b.py hddata_b.csv > hddata_c.csv
+	rm hddata_b.csv
+fi
 
-echo
-date
-echo "Running pre-process ..."
-./preprocess.py hddata_c.csv > hddata_d.csv
-#rm hddata_c.csv
+if [ -e "hddata_c.csv" ]; then
+	echo
+	date
+	echo "Running pre-process (regular expresion processing) ..."
+	./preprocess.py hddata_c.csv > hddata_d.csv
+	#rm hddata_c.csv
+fi
 
-echo
-date
-echo "Generate description/solution files ..."
-./util_csvcat.py hddata_d.csv 13 > hddata_d_des.csv
-./util_csvcat.py hddata_d.csv 14 > hddata_d_sol.csv
-#rm hddata_d.csv
+if [ -e "hddata_d.csv" ]; then
+	echo
+	date
+	echo "Running process (cascaded finite-state transducer) ..."
+	./process.py hddata_d.csv 13 > hddata_e.csv
+	#rm hddata_d.csv
+fi
